refactor(routes): moved routes to app.js
