+++ server.js
@@ -5,7 +5,9 @@ const toasters = require('./routes/toasters');
 const customers = require('./routes/customers');
 const auth = require('./routes/auth');
 var BASE_API_PATH = "/api/v1";
-
+const router = express.Router();
+const circuitBreaker = require("./circuitBreaker");
+const dashboard = require('hystrix-dashboard')
 console.log("Starting API server...");
 
 var app = express();
@@ -19,5 +21,17 @@ app.get("/", (req, res) => {
 app.use(BASE_API_PATH + '/customers', customers);
 app.use(BASE_API_PATH + '/toasters', toasters);
 app.use(BASE_API_PATH + '/auth', auth);
+
+app.use(router);
+
+app.use(
+    dashboard({
+        idleTimeout: 4000,
+        interval: 2000,
+        proxy: true,
+    })
+);
+circuitBreaker.initHystrixStream(router);
+
 
 module.exports = app;
